% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imc_plot.R
\name{create_imc_plot}
\alias{create_imc_plot}
\title{Create Immune Feature QC Plot}
\arguments{
\item{imc_df}{A data frame containing the immune cell data.}

\item{project_name}{A string representing the name of the project, used as the plot title.}

\item{IMC_percent}{A numeric value representing the percentage of immune cells.}
}
\value{
A ggplot object representing the immune feature QC plot.
}
\description{
Helper function to generate scatter plot of UMI & gene counts per cell barcode
with points colored based on whether they are immune cells (IMC) or non-immune cells.
}
\keyword{internal}
