% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_grid.R
\name{create_plot_grid}
\alias{create_plot_grid}
\title{Save a Grid of Plots}
\usage{
create_plot_grid(plots, file_path, nrow, ncol = 5)
}
\arguments{
\item{plots}{A list of ggplot objects to be arranged in a grid.}

\item{file_path}{A string representing the file path where the plot grid will be saved.}

\item{nrow}{An integer representing the number of rows in the plot grid. Default is 3.}

\item{ncol}{An integer representing the number of columns in the plot grid. Default is 5.}
}
\value{
None. The function saves the plot grid to the specified file path.
}
\description{
Helper function to save a grid of plots with a specified number of columns and rows.
}
\keyword{internal}
