% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rbc_plot.R
\name{create_rbc_plot}
\alias{create_rbc_plot}
\title{Create RBC Plot}
\arguments{
\item{rbc_df}{A data frame containing the RBC data.}

\item{project_name}{A string representing the name of the project, used as the plot title.}

\item{RBC_percent}{A numeric value representing the percentage of RBC contamination.}

\item{initial_cells}{An integer representing the initial number of cells.}
}
\value{
A 'ggplot2' object representing the RBC plot.
}
\description{
This helper function generates a scatter plot showing the relationship
between haemoglobin expression and CD45 expression, with points colored based
on whether they are RBCs or non-RBCs.
}
\examples{
\dontrun{
# Create test dataset
rbc_data <- data.frame(
  hemo.percent = runif(100, min = 0, max = 100),
  ptprc.percent = runif(100, min = 0, max = 100),
  RBC = sample(c("RBC", "non-RBC"), 100, replace = TRUE)
)

create_rbc_plot(rbc_df = rbc_data,
                project_name = "test",
                RBC_percent = 15,
                initial_cells = 1000)

}
}
\keyword{internal}
