% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DamageDetective_core.R
\name{DamageDetective_core}
\alias{DamageDetective_core}
\title{DamageDetective_core}
\arguments{
\item{project_name}{String with project or sample name}

\item{filtered_path}{Directory of filtered alignment output}

\item{seurat_input}{Seurat object to be used as input over raw files. Default NULL}

\item{min_cells}{In how many cells should a gene be expressed to be kept}

\item{soupx}{Perform ambient RNA correction, if TRUE raw_path must be given. Default is FALSE}

\item{raw_path}{Directory of unfiltered alignment output}

\item{filter_rbc}{Whether or not red blood cells should be removed. Default is TRUE}

\item{hemo_threshold}{Percent hemoglobin expression above which cells are filtered. Default is 50}

\item{isolate_cd45}{Discard non-immune cells. Default is FALSE}

\item{filter_output}{Should output contain no damaged cells. Default is TRUE}

\item{output_path}{Directory where output can be generated}

\item{resolution}{Numeric between 0 and 1.6 describing cluster division. Default 1}

\item{cluster_ranks}{Numeric describing the number of top ranking clusters to be included as damaged cells. Default 1.}

\item{organism}{"Hsap" if human sample or "Mmus" if mouse sample}

\item{verbose}{Print messages to the console. Default is TRUE}
}
\value{
(list) Output list storing the final Seurat object
}
\description{
Core helper function enabling the DamageDetective modules to run
}
\examples{
\donttest{
# Load test data
data("test_data", package = "DamageDetective")

test <- DamageDetective(
  project_name  = "test",
  seurat_input  = test_data,
  filter_rbc    = FALSE,
  output_path   = tempdir()
)

}

}
\keyword{internal}
