% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_annotations.R
\docType{data}
\name{human_annotations}
\alias{human_annotations}
\title{Annotations}
\format{
An object of class \code{data.frame} with 67667 rows and 3 columns.
}
\source{
AnnotationHub
}
\usage{
human_annotations
}
\description{
Gene annotations for human (Homo sapiens).
}
\examples{
data(human_annotations)
}
\keyword{datasets}
