% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_damage_cells.R
\name{identify_damage_cells}
\alias{identify_damage_cells}
\title{Identify damage cells}
\arguments{
\item{nf_umi_ed_ct}{data frame, with four columns. The first three columns
should match the output from the \code{identify_empty_drops} function. The
fourth column should contain cell type names.}

\item{nf_sep}{numeric, the minimum separation of the nuclear fraction score
required between the cell and damage cell populations}

\item{umi_sep_perc}{numeric, this is the minimum percentage of UMIs which the
damaged cell population is required to have compared to the cell
population. For example, if the mean UMI of the distribution fit to the
whole cell population is 10,000 UMIs, the mean of the distribution fit to
the damaged cell population must be at less than 7,000 UMIs if the umi_sep
parameter is 30 (\%)}

\item{verbose}{logical, whether to print updates and progress while fitting
with EM}
}
\value{
list, of length two. The first element in the list contains a data
frame with the same dimensions input to the \code{nf_umi_ed_ct} argument, with
"damaged_cell" now recorded in the third column.
}
\description{
This function uses a combination of the cell UMI counts and the
nuclear fraction score to assign each cell one of two values; "cell" or
"damaged_cell". This is based on the idea that damaged cells have a lower
UMI count and higher nuclear fraction than whole cells. The expected input
is a data frame with four columns. The first three columns should contain;
the nuclear fraction score, total UMIs and a character vector describing
each cell as "cell" or "empty_droplet". This is the format output by the
\code{identify_empty_drops} function. The fourth column should be a character
vector with user-assigned cell types. Internally, the provided data frame
is split by cell type and a Gaussian mixture model with a maximum of two
components is fit to the umi counts and nuclear fraction scores. The
parameters of the model are estimated using expectation maximisation (EM)
with the \code{mclust} package. The best model is selected using the Bayesian
Information Criterion (BIC). The two populations (cells and damaged cells)
are assumed to have equal variance (mclust model name "EEI").
EM Model: It’s used to find and separate multiple groups or clusters within
the data. It’s not just about finding points that don’t fit in; it’s about
identifying distinct groups that may exist in the data.
}
\keyword{internal}
