% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_empty_droplets.R
\name{identify_empty_droplets}
\alias{identify_empty_droplets}
\title{Identify empty droplets}
\arguments{
\item{nf_umi}{data frame, containing two columns; the nuclear fraction
estimates in the first column and the total UMI count for each barcode in
the second column}

\item{nf_rescue}{numeric, a rescue parameter defining a minimum nuclear
fraction score between zero and one. This is used in combination with
\code{umi_rescue} to identify cells that were misidentified as empty droplets}

\item{umi_rescue}{integer, a rescue parameter defining a minimum UMI count.
This is used in combination with \code{nf_rescue} to identify cells that were
misidentified as empty droplets}

\item{plot_name}{character, if provided a plot will be saved with the
provided name}

\item{plot_path}{character, if provided a plot will be saved to the provided
path}

\item{plot_width}{numeric, plot width in cm}

\item{plot_height}{numeric, plot height in cm}

\item{pdf_png}{character, either "png" or "pdf"}
}
\value{
data frame, the original data frame is returned plus an additional
column identifying each barcode as a "cell" or "empty_droplet"
}
\description{
Identify a suitable nuclear fraction
cut-off point to guide the identification of empty droplets. To do this it
calculates the kernel density estimate of the input nuclear fraction scores
and identifies the trough after the first peak, which is assumed to
represent the population of empty droplets.
}
\examples{
\dontrun{
# Assuming `nf_umi` is a data frame with the required structure.
nf_umi <- data.frame(
  nf = runif(100, min = 0, max = 1),
  umi = sample(100:10000, 100, replace = TRUE)
)
nf_rescue <- 0.05
umi_rescue <- 1000
plot_name <- "empty_droplets_plot"
plot_path <- "path/to/save/plot"
plot_width <- 18
plot_height <- 13
pdf_png <- "png"

result <- identify_empty_droplets(nf_umi,
                                  nf_rescue,
                                  umi_rescue,
                                  plot_name,
                                  plot_path,
                                  plot_width,
                                  plot_height,
                                  pdf_png)
print(result)
}
}
\keyword{internal}
