% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiric_calculation.R
\name{limiric_calculation}
\alias{limiric_calculation}
\title{Identify damaged cells with limiric}
\arguments{
\item{organism}{A string representing the organism ("Hsap", "Mmus").}

\item{Seurat}{A 'Seurat' object containing the single-cell RNA-seq data.}

\item{resolution}{Numeric between 0 and 1.6 describing cluster division. Default is 1.}

\item{cluster_ranks}{Numeric describing the number of top ranking clusters to be included as damaged cells. Default 1.}

\item{annotations}{A data frame containing gene annotations.}

\item{initial_cells}{An integer representing the initial number of cells.}

\item{project_name}{A string representing the name of the project, used for plot titles.}

\item{output_path}{A string representing the path to save the output plots.}
}
\value{
A list containing the updated 'Seurat' object and the reduced 'Seurat' object.
}
\description{
This helper function calculates the 'limiric' metrics for a given 'Seurat' object and organism.
It identifies mitochondrial and ribosomal genes, reduces the 'Seurat' object based on these genes,
and calculates various metrics including mitochondrial and ribosomal expression complexity.
}
\examples{
\donttest{
# Load test data
data("test_data", package = "limiric")
data("human_annotations", package = "limiric")

result <- limiric_calculation(organism = "Hsap",
                              Seurat = test_data,
                              annotations = human_annotations,
                              resolution = 0.6,
                              cluster_ranks = 1,
                              initial_cells = 1000,
                              project_name = "Test",
                              output_path = tempdir())
print(result)
}
}
\keyword{internal}
