% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiric_core.R
\name{limiric_core}
\alias{limiric_core}
\title{limiric_core}
\arguments{
\item{project_name}{String with project or sample name}

\item{filtered_path}{Directory of filtered alignment output}

\item{seurat_input}{Seurat object to be used as input over raw files. Default NULL}

\item{min_cells}{In how many cells should a gene be expressed to be kept}

\item{soupx}{Perform ambient RNA correction, if TRUE raw_path must be given. Default is FALSE}

\item{raw_path}{Directory of unfiltered alignment output}

\item{droplet_qc}{Verify output with droplet_qc, if TRUE velocyto_path must be given. Default is FALSE}

\item{velocyto_path}{Directory of Veocyto filtered alignment output}

\item{filter_rbc}{Whether or not red blood cells should be removed, Default is TRUE}

\item{isolate_cd45}{Discard non-immune cells. Default is FALSE}

\item{filter_output}{Should output contain no damaged cells. Default is TRUE}

\item{output_path}{Directory where limiric output cen be generated}

\item{organism}{"Hsap" if human sample or "Mmus" if mouse sample}
}
\value{
(list) Output list storing the final Seurat object
}
\description{
Core helper function enabling the limiric modules to run
}
\examples{
\dontrun{

# Detect damaged cells

SRR1234567 <- limiric_core(
  project_name  = "SRR1234567",
  filtered_path = "/home/user/alignment/SRR1234567/filtered/",
  soupx        = TRUE,
  raw_path      = "/home/user/alignment/SRR1234567/raw/",
  droplet_qc    = TRUE,
  isolate_cd45  = TRUE,
  velocyto_path = "/home/user/alignment/velocyto/",
  output_path   = "/home/user/alignment/limiric/"
)

}

}
\keyword{internal}
