% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiricCore.R
\name{limiricCore}
\alias{limiricCore}
\title{limiricCore}
\arguments{
\item{ProjectName}{String with project or sample name}

\item{FilteredPath}{Directory of filtered alignment output}

\item{SeuratInput}{Seurat object to be used as input over raw files. Default NULL}

\item{MinCells}{In how many cells should a gene be expressed to be kept}

\item{SoupX}{Perform ambient RNA correction, if TRUE RawPath must be given. Default is FALSE}

\item{RawPath}{Directory of unfiltered alignment output}

\item{DropletQC}{Verify output with DropletQC, if TRUE VelocytoPath must be given. Default is FALSE}

\item{VelocytoPath}{Directory of Veocyto filtered alignment output}

\item{FilterRBC}{Whether or not red blood cells should be removed, Default is TRUE}

\item{IsolateCD45}{Discard non-immune cells. Default is FALSE}

\item{FilterOutput}{Should output contain no damaged cells. Default is TRUE}

\item{OutputPath}{Directory where limiric output cen be generated}

\item{Organism}{"Hsap" if human sample or "Mmus" if mouse sample}
}
\value{
(list) Output list storing the final Seurat object
}
\description{
Helper function to enable the limiric function to have list inputs
}
\examples{
\dontrun{
# Detect damaged cells

SRR1234567 <- limiricCore(
  ProjectName  = "SRR1234567",
  FilteredPath = "/home/user/alignment/SRR1234567/filtered/",
  SoupX        = TRUE,
  RawPath      = "/home/user/alignment/SRR1234567/raw/",
  DropletQC    = TRUE,
  IsolateCD45  = TRUE,
  VelocytoPath = "/home/user/alignment/velocyto/",
  OutputPath   = "/home/user/alignment/limiric/"
)
}
}
