% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_annotations.R
\docType{data}
\name{mouse_annotations}
\alias{mouse_annotations}
\title{Mouse Annotations}
\format{
A data frame with 57,180 rows and 3 variables:
\describe{
\item{gene_name}{Character. The name of the gene.}
\item{gene_biotype}{Character. The biotype of the gene.}
\item{description}{Character. A description of the gene.}
}
}
\source{
AnnotationHub
}
\usage{
mouse_annotations
}
\description{
This dataset contains annotations for the mouse genome.
}
\examples{
data(mouse_annotations)
head(mouse_annotations)
}
\keyword{datasets}
