% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_rb_genes.R
\docType{data}
\name{mt_rb_genes}
\alias{mt_rb_genes}
\title{Human Mitochondrial and Ribosomal Genes}
\format{
A character vector with 201 entries
}
\source{
AnnotationHub
}
\usage{
mt_rb_genes
}
\description{
Annotations for the mitochondrial and ribosomal genes within the human genome.
}
\examples{
data(mt_rb_genes)
head(mt_rb_genes)
}
\keyword{datasets}
