% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{Test Data Seurat Object}
\format{
A Seurat object with 10 009 features and 500 cells.
\describe{
\item{features}{Numeric. The number of features (genes) in the dataset.}
\item{cells}{Numeric. The number of cells in the dataset.}
}
}
\source{
Generated for testing purposes
}
\usage{
test_data
}
\description{
This dataset contains a Seurat object for testing purposes.
}
\examples{
data(test_data)
test_data
}
\keyword{datasets}
