% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_EM.R
\name{assess_EM}
\alias{assess_EM}
\title{assess_EM}
\arguments{
\item{em}{Mclust, result of EM on log10(UMI counts) and nf to assess
the likelihood of two cell populations (cells and damaged cells) or just cells.}

\item{umi_thresh}{numeric, percentage by which damaged_cell UMI counts
must be below the cell distribution mean to be classified as damaged cells
(e.g., if umi_thresh = 30, damaged_cell UMI mean must be < 70\% of cell mean,
otherwise all barcodes are returned as cells).}

\item{nf_thresh}{numeric, minimum nuclear fraction difference required
between cell and damaged_cell distributions to classify as damaged cells.}
}
\value{
data frame, with three columns containing the log10(UMI counts),
nuclear fraction score and the assigned cell status; "cell" or
"damaged_cell"
}
\description{
powellgenomicslab/DropletQC helper function called by
\code{identify_damaged_cells} not intended for more general use.
The function is designed to detect and label damaged cells based on
their UMI counts and nuclear fraction, relying on the EM model
to differentiate between potential cell populations. The thresholds
(umi_thresh and nf_thresh) ensure that only cells with significant
differences in these metrics are considered damaged.
}
\details{
Assess Expectation-Maximization model results to assign cells as damaged or not
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Example usage:
# Assuming `em_result` is the result of an EM model on log10(UMI counts) and nf
# and you have set thresholds for UMI and nuclear fraction.
em_result <- Mclust(data = your_data)  # Replace `your_data` with actual data
umi_threshold <- 30
nf_threshold <- 0.1
result <- assess_EM(em_result, umi_threshold, nf_threshold)
print(result)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
