% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_dropletqc_plot.R
\name{create_dropletqc_plot}
\alias{create_dropletqc_plot}
\title{Create DropletQC plot output}
\arguments{
\item{Seurat}{A 'Seurat' object containing the single-cell RNA-seq data.}

\item{limiric}{A 'Seurat' object containing the 'limiric' data.}

\item{output_path}{A string representing the path to save the plots.}

\item{project_name}{A string representing the name of the project, used as the plot title.}

\item{damaged_percent}{A numeric value representing the percentage of damaged cells.}

\item{initial_cells}{An integer representing the initial number of cells.}
}
\value{
None. The function saves the plots as PNG files.
}
\description{
This helper function generates a cluster and scatter plot
for viewing labels in reduced space and scatter plot of nf vs UMI_count.
The result is saved as a PNG.
}
\examples{
\dontrun{
# Assuming `seurat_obj` is a Seurat object
# and `limiric_obj` is a Seurat object

# Define inputs
seurat_obj <- CreateSeuratObject(counts = your_counts)
limiric_obj <- CreateSeuratObject(counts = your_lim_counts)
output_path <- "path/to/save/plots"
project_name <- "MyProject"
damaged_percent <- 5.0
initial_cells <- 1000

create_dropletqc_plot(seurat_obj,
                      limiric_obj,
                      output_path,
                      project_name,
                      damaged_percent,
                      initial_cells)
}
}
\keyword{internal}
