% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_imc_plot.R
\name{create_imc_plot}
\alias{create_imc_plot}
\title{Create Immune Feature QC Plot}
\arguments{
\item{imc_df}{A data frame containing the immune cell data.}

\item{project_name}{A string representing the name of the project, used as the plot title.}

\item{IMC_percent}{A numeric value representing the percentage of immune cells.}
}
\value{
A ggplot object representing the immune feature QC plot.
}
\description{
Helper function to generate scatter plot of UMI & gene counts per cell barcode
with points colored based on whether they are immune cells (IMC) or non-immune cells.
}
\examples{
\donttest{
# Create test dataset
imc_data <- data.frame(
  nCount_RNA   = runif(100, min = 500, max = 10000),
  nFeature_RNA = runif(100, min = 200, max = 3000),
  IMC = sample(c("IMC", "non-IMC"), 100, replace = TRUE)
)

create_imc_plot(imc_df = imc_data,
                project_name = "test",
                IMC_percent = 25)


}
}
\keyword{internal}
