% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_rbc_plot.R
\name{create_rbc_plot}
\alias{create_rbc_plot}
\title{Create RBC Plot}
\arguments{
\item{rbc_df}{A data frame containing the RBC data.}

\item{project_name}{A string representing the name of the project, used as the plot title.}

\item{RBC_percent}{A numeric value representing the percentage of RBC contamination.}

\item{initial_cells}{An integer representing the initial number of cells.}
}
\value{
A ggplot object representing the RBC plot.
}
\description{
This helper function generates a scatter plot showing the relationship between haemoglobin expression and CD45 expression,
with points colored based on whether they are RBCs or non-RBCs.
}
\keyword{internal}
