% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropletqc_calculation.R
\name{dropletqc_calculation}
\alias{dropletqc_calculation}
\title{Identify Empty Droplets and Damaged Cells}
\arguments{
\item{Seurat}{A 'Seurat' object containing the single-cell RNA-seq data.}
}
\value{
The updated 'Seurat' object with 'DropletQC' results added to its meta data.
}
\description{
This helper function extracts nf meta data and associated cell barcodes from a 'Seurat' object,
identifies empty droplets using identify_empty_droplets() and identifies
damaged cells using the identify_damage_cells() function. The results are added to the
'Seurat' object's meta data.
}
\examples{
\dontrun{
# Example usage:
# Assuming `seurat_obj` is a Seurat object containing the single-cell RNA-seq data.
seurat_obj <- CreateSeuratObject(counts = your_counts)  # Replace `your_counts` with actual data
updated_seurat_obj <- dropletqc_calculation(seurat_obj)
print(updated_seurat_obj)
}
}
\keyword{internal}
