% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/human_annotations.R
\docType{data}
\name{human_annotations}
\alias{human_annotations}
\title{Human Annotations}
\format{
A data frame with 72,035 rows and 3 variables:
\describe{
\item{gene_name}{Character. The name of the gene.}
\item{gene_biotype}{Character. The biotype of the gene.}
\item{description}{Character. A description of the gene.}
}
}
\source{
AnnotationHub
}
\usage{
human_annotations
}
\description{
This dataset contains annotations for the human genome.
}
\examples{
data(human_annotations)
head(human_annotations)
}
\keyword{datasets}
