% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiric_calculation.R
\name{limiric_calculation}
\alias{limiric_calculation}
\title{Identify damaged cells with limiric}
\arguments{
\item{organism}{A string representing the organism ("Hsap", "Mmus").}

\item{Seurat}{A Seurat object containing the single-cell RNA-seq data.}

\item{annotations}{A data frame containing gene annotations.}

\item{initial_cells}{An integer representing the initial number of cells.}

\item{project_name}{A string representing the name of the project, used for plot titles.}

\item{output_path}{A string representing the path to save the output plots.}
}
\value{
A list containing the updated Seurat object and the reduced Seurat object (limiric).
}
\description{
This helper function calculates the LIMIRIC metrics for a given Seurat object and organism.
It identifies mitochondrial and ribosomal genes, reduces the Seurat object based on these genes,
and calculates various metrics including mitochondrial and ribosomal expression, and complexity.
}
\keyword{internal}
