% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiric_calculation.R
\name{limiric_calculation}
\alias{limiric_calculation}
\title{Identify damaged cells with limiric}
\arguments{
\item{organism}{A string representing the organism ("Hsap", "Mmus").}

\item{Seurat}{A Seurat object containing the single-cell RNA-seq data.}

\item{annotations}{A data frame containing gene annotations.}

\item{initial_cells}{An integer representing the initial number of cells.}

\item{project_name}{A string representing the name of the project, used for plot titles.}

\item{output_path}{A string representing the path to save the output plots.}
}
\value{
A list containing the updated Seurat object and the reduced Seurat object (limiric).
}
\description{
This helper function calculates the LIMIRIC metrics for a given Seurat object and organism.
It identifies mitochondrial and ribosomal genes, reduces the Seurat object based on these genes,
and calculates various metrics including mitochondrial and ribosomal expression, and complexity.
}
\examples{
\dontrun{
# Assuming `seurat_obj` is a Seurat object
seurat_obj <- CreateSeuratObject(counts = your_counts)
annotations <- data.frame(
  gene_name = c("MT-CO1", "RPS3", "RPL5"),
  gene_biotype = c("protein_coding", "protein_coding", "protein_coding"))
organism <- "Hsap"
initial_cells <- 1000
project_name <- "MyProject"
output_path <- "path/to/save/plots"

result <- limiric_calculation(organism,
                              seurat_obj,
                              annotations,
                              initial_cells,
                              project_name,
                              output_path)
print(result)
}
}
\keyword{internal}
