% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiricCore.R
\name{limiricCore}
\alias{limiricCore}
\title{limiricCore}
\arguments{
\item{ProjectName}{(Character) Name of the project or sample being analysed}

\item{FilteredPath}{(Character) Directory of filtered, gzipped, alignment output containing matrix.mtx.gz, features.tsv.gz, and barcodes.tsv.gz}

\item{SeuratInput}{(object) Seurat object to be used as input}

\item{MinCells}{(numeric) In how many cells should a gene be expressed to be retained in the count matrix}

\item{SoupX}{(logical) Whether or not ambient RNA correction should be performed, if TRUE RawPath must be given}

\item{RawPath}{(character) Directory of unfiltered, gzipped, alignment output containing matrix.mtx.gz, features.tsv.gz, and barcodes.tsv.gz}

\item{DropletQC}{(logical) Whether or not limiric annotations should be intersected with DropletQC damaged cell annotations, if TRUE VelocytoPath must be given}

\item{VelocytoPath}{(character) Directory of Veocyto filtered, gzipped, alignment output containing spliced.mtx.gz, unspliced.mtx.gz, features.tsv.gz, and barcodes.tsv.gz}

\item{FilterRBC}{(logical) Whether or not red blood cells should be removed}

\item{IsolateCD45}{(logical) Whether or not immune cells should be isolated}

\item{FilterOutput}{(logical) Whether or not output should filter out damaged cells}

\item{OutputPath}{(character) Directory where limiric output is generated}

\item{Organism}{(character) Either "Hsap" if human sample or "Mmus" if mouse}
}
\value{
(list) Output list storing the final Seurat object
}
\description{
Helper function to enable the limiric function to have list inputs
}
\examples{
\dontrun{
# Detect damaged cells

SRR1234567 <- limiricCore(
  ProjectName  = "SRR1234567",
  FilteredPath = "/home/user/alignment/SRR1234567/filtered/",
  SoupX        = TRUE,
  RawPath      = "/home/user/alignment/SRR1234567/raw/",
  DropletQC    = TRUE,
  IsolateCD45  = TRUE,
  VelocytoPath = "/home/user/alignment/velocyto/",
  OutputPath   = "/home/user/alignment/limiric/"
)
}
}
