% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse_annotations.R
\docType{data}
\name{mouse_annotations}
\alias{mouse_annotations}
\title{Annotations}
\format{
An object of class \code{data.frame} with 57180 rows and 3 columns.
}
\source{
AnnotationHub
}
\usage{
mouse_annotations
}
\description{
Gene annotations for mouse (Mus musculus).
}
\examples{
data(mouse_annotations)
}
\keyword{datasets}
