% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soupx_calculation.R
\name{soupx_calculation}
\alias{soupx_calculation}
\title{Perform SoupX Correction}
\arguments{
\item{raw_path}{A string representing the path to the raw data.}

\item{table_of_counts}{A matrix of counts.}

\item{min_cells}{An integer representing the minimum number of cells.}

\item{project_name}{A string representing the name of the project.}
}
\value{
A Seurat object with SoupX corrected counts
}
\description{
This helper function performs SoupX correction with the raw output (must be zipped), estimates the contamination,
and outputs the results in a Seurat object for continued workflow.
}
\keyword{internal}
