% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soupx_calculation.R
\name{soupx_calculation}
\alias{soupx_calculation}
\title{Perform SoupX Correction}
\arguments{
\item{table_of_droplets}{A matrix of counts from the unfiltered alignment output.}

\item{table_of_counts}{A matrix of counts from the filtered alignment output.}

\item{min_cells}{An integer representing the minimum number of cells.}

\item{project_name}{A string representing the name of the project.}
}
\value{
A Seurat object with SoupX corrected counts
}
\description{
This helper function performs SoupX correction with the raw output (must be zipped), estimates the contamination,
and outputs the results in a Seurat object for continued workflow.
}
\examples{
\donttest{
# Load test data from Seurat
tod = Seurat::Read10X(system.file('extdata','toyData','raw_gene_bc_matrices',
'GRCh38', package='SoupX'))
toc = Seurat::Read10X(system.file('extdata','toyData','filtered_gene_bc_matrices',
'GRCh38', package='SoupX'))

result <- soupx_calculation(table_of_droplets = tod,
                            table_of_counts = toc,
                            min_cells = 0,
                            project_name = "test")
}
}
\keyword{internal}
